requirements.txt
numpy>=1.24.0,<2.0.0
xarray>=2023.1.0
zarr>=2.14.2
s3fs==2023.9.0
setuptools>=65.5.0
pytz>=2023.3
treelib>=1.2.5
tqdm>=4.60.0
timm>=0.9.12
hydra-core>=1.2.0
termcolor>=2.1.1
wandb>=0.13.7
mlflow>=2.1.1
pytest>=6.0.0
pyyaml>=6.0
h5py>=3.7.0
netcdf4>=1.6.3
ruamel.yaml>=0.17.21
scikit-learn>=1.2.2,<=1.6.0
scikit-image>=0.24.0
vtk>=9.2.6,<9.4.0
pyvista>=0.40.1
cftime>=1.6.2
einops>=0.7.0
shapely>=2.0.6
onnx>=1.14.0
pandas>=2.2.2
omegaconf>=2.3.0
dask>=2024.11.1
mpi4py
seaborn
torchdata<=0.9.0
openfermion==1.5.1
pybind11
torchmetrics
e3nn==0.4.4
deepxde
ase
rdkit
xtb
pymatgen
torch-runstats>=0.2.0
torch-ema>=0.3.0
matscipy
python-hostlist
configargparse
opt_einsum
prettytable
lmdb
orjson
matplotlib
contextlib2
ml-collections==0.1.1
dm-tree==0.1.8
dm-haiku==0.0.14
diffrax==0.6.0
biopandas==0.5.1
biopython==1.84
pyrsistent